+++ qpainterfilm.h
@@ -8,14 +8,14 @@ class QPainterFilm :
     public Film
 {
 public:
-    QPainterFilm(QPainter &qPainter, QPen &qPen, int width, int height);
-    void write(int xPos, int yPos, int zDepth, int r, int g, int b) override;
+    QPainterFilm(QPainter &qPainter, QPen &qPen, uint32_t width, uint32_t height);
+    void write(uint32_t xPos, uint32_t yPos, float zDepth, float r, float g, float b) override;
 
 private:
     QPainter &qPainter;
     QPen &qPen;
-    int const width, const height;
-    std::vector<int> depthBuffer;
+    uint32_t const width, const height;
+    std::vector<float> depthBuffer;
 };
 
 #endif QPAINTERFILM_H
